Revamped workflow, striping away some idiosincrasies
Most notably, got rid of the previous verions of  'attached state'
which made a copy of the Workflowstate object and put in a reference
to a workflow instance in the copy.

This version uses descriptors, and a simple proxy object instead.

For the users of workflows, a new way to declare transitions
that don't require any code is simply declaring

class BlaWorkflow(Workflow):
    state1 = WorkflowState(...)
    my_transition = state1.transition(...)

And voilá - my_transition is a callable that will
make the state transition.