Implements declarative Permissions
Permissions can now be used in a declarative form in
the Workflow class body, and bound to state and or roles
automatically.

Workflows permissions were also promoted
to descriptors, so that when retriveing a permission attribute from a
workflow instance, its truth value is returned.

Check tests for examples of the new ways to declare permissions